fix: ensure sqs fifo tests correct

+++ tests/Paramore.Brighter.AWS.Tests/MessagingGateway/Sqs/Fifo/Proactor/When_queues_missing_verify_throws_async.cs
@@ -9,13 +9,13 @@ using Xunit;
 namespace Paramore.Brighter.AWS.Tests.MessagingGateway.Sqs.Fifo.Proactor;
 
 [Trait("Category", "AWS")]
-public class AWSValidateQueuesTestsAsync : IAsyncDisposable
+public class AwsValidateQueuesTestsAsync : IAsyncDisposable
 {
     private readonly AWSMessagingGatewayConnection _awsConnection;
     private readonly SqsSubscription<MyCommand> _subscription;
-    private ChannelFactory _channelFactory;
+    private ChannelFactory? _channelFactory;
 
-    public AWSValidateQueuesTestsAsync()
+    public AwsValidateQueuesTestsAsync()
     {
         var subscriptionName = $"Producer-Send-Tests-{Guid.NewGuid().ToString()}".Truncate(45);
         var queueName = $"Producer-Send-Tests-{Guid.NewGuid().ToString()}".Truncate(45);
@@ -24,7 +24,11 @@ public class AWSValidateQueuesTestsAsync : IAsyncDisposable
         _subscription = new SqsSubscription<MyCommand>(
             subscriptionName: new SubscriptionName(subscriptionName),
             channelName: new ChannelName(queueName),
-            channelType: ChannelType.PointToPoint, routingKey: routingKey, queueAttributes: new SqsAttributes(type: SqsType.Fifo), makeChannels: OnMissingChannel.Validate);
+            channelType: ChannelType.PointToPoint, 
+            routingKey: routingKey,
+            messagePumpType: MessagePumpType.Proactor,
+            queueAttributes: new SqsAttributes(type: SqsType.Fifo), 
+            makeChannels: OnMissingChannel.Validate);
 
         _awsConnection = GatewayFactory.CreateFactory();
     }
@@ -41,6 +45,7 @@ public class AWSValidateQueuesTestsAsync : IAsyncDisposable
 
     public async ValueTask DisposeAsync()
     {
-        await _channelFactory.DeleteTopicAsync();
+        if (_channelFactory != null)
+            await _channelFactory.DeleteTopicAsync();
     }
 }
